requirements.txt
annotated-types==0.6.0
anyio==4.3.0
backports.tarfile==1.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dnspython==2.6.1
docutils==0.21.2
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.3
fastapi-namespace==0.1.0
fastapi-redis-vet1ments==0.2.6
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jeepney==0.8.0
Jinja2==3.1.4
keyring==25.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
nh3==0.2.17
orjson==3.10.3
pkginfo==1.10.0
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
readme_renderer==43.0
redis==5.0.4
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
twine==5.0.0
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
zipp==3.18.1
